Scripts/requirements.txt
Pillow==9.0.1
PyYAML==6.0